requirements.txt
asgiref==3.8.1
Django==5.1.3
django-auth-ldap==5.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-ldap==3.4.4
sqlparse==0.5.2
typing_extensions==4.12.2
